Add number of conditionals in computation of priority score

+++ soot-infoflow/src/soot/jimple/infoflow/results/InfoflowResults.java
@@ -595,7 +595,7 @@ public class InfoflowResults {
 		}
 
 		// Sort by length of taint paths
-		Collections.sort(prioritizedResults, Comparator.comparingInt(DataFlowResult::getPriorityScore));
+		Collections.sort(prioritizedResults, Comparator.comparingDouble(DataFlowResult::getPriorityScore));
 		return prioritizedResults;
 	}
 }
